+++ src/main/java/com/Chomusuke/logic/Transaction.java
@@ -15,7 +15,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.chomusuke.transactions;
+package com.chomusuke.logic;
 
 /**
  * Provides memory storage of transactions.
